requirements.txt
pyelftools==0.24
PyYAML==3.12
spread-confiture==0.17.3rc1
